+++ package.json
@@ -15,6 +15,7 @@
     "moment-timezone": "^0.5.37",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.0"
